# Breakup cops cursor logic; simplify render