--- a/internal/cops/display/cursor.go
+++ b/internal/cops/display/cursor.go
@@ -170,6 +170,13 @@ func (c Cursor) up(buf []byte, n int) ([]byte, Cursor) {
 	return buf, c
 }
 
+func (c Cursor) right(buf []byte, n int) ([]byte, Cursor) {
+	buf = append(buf, "\033["...)
+	buf = strconv.AppendInt(buf, int64(n), 10)
+	buf = append(buf, "C"...)
+	return buf, c
+}
+
 // Go moves the cursor to another position, preferring to use relative motion,
 // using line relative if the column is unknown, using display origin relative
 // only if the line is also unknown. If the column is unknown, use "\r" to seek
@@ -199,15 +206,12 @@ func (c Cursor) Go(buf []byte, to image.Point) ([]byte, Cursor) {
 		buf, c = c.up(buf, -n)
 	}
 
-	// LEFT OR RIGHT
-	if to.X < c.Position.X {
+	if n := to.X - c.Position.X; n > 0 {
+		buf, c = c.right(buf, n)
+	} else if to.X < c.Position.X {
 		buf = append(buf, "\033["...)
 		buf = strconv.AppendInt(buf, int64(c.Position.X-to.X), 10)
 		buf = append(buf, "D"...)
-	} else if to.X > c.Position.X {
-		buf = append(buf, "\033["...)
-		buf = strconv.AppendInt(buf, int64(to.X-c.Position.X), 10)
-		buf = append(buf, "C"...)
 	}
 
 	c.Position = to