--- a/internal/cops/display/cursor.go
+++ b/internal/cops/display/cursor.go
@@ -167,6 +167,7 @@ func (c Cursor) up(buf []byte, n int) ([]byte, Cursor) {
 	buf = append(buf, "\033["...)
 	buf = strconv.AppendInt(buf, int64(n), 10)
 	buf = append(buf, "A"...)
+	c.Position.Y -= n
 	return buf, c
 }
 
@@ -174,6 +175,7 @@ func (c Cursor) left(buf []byte, n int) ([]byte, Cursor) {
 	buf = append(buf, "\033["...)
 	buf = strconv.AppendInt(buf, int64(n), 10)
 	buf = append(buf, "D"...)
+	c.Position.X -= n
 	return buf, c
 }
 
@@ -181,6 +183,7 @@ func (c Cursor) right(buf []byte, n int) ([]byte, Cursor) {
 	buf = append(buf, "\033["...)
 	buf = strconv.AppendInt(buf, int64(n), 10)
 	buf = append(buf, "C"...)
+	c.Position.X += n
 	return buf, c
 }
 
@@ -219,7 +222,6 @@ func (c Cursor) Go(buf []byte, to image.Point) ([]byte, Cursor) {
 		buf, c = c.left(buf, -n)
 	}
 
-	c.Position = to
 	return buf, c
 }