More internal surface area, used by render.