@jcorbin 2017-12-09
Small perf win, if any:
```                                                      
name                old time/op    new time/op    delta                 
_displayDemo/4-4      4.57µs ± 0%    4.16µs ± 0%   ~     (p=1.000 n=1+1)         
_displayDemo/8-4      16.2µs ± 0%    15.0µs ± 0%   ~     (p=1.000 n=1+1)   
_displayDemo/16-4     61.2µs ± 0%    58.0µs ± 0%   ~     (p=1.000 n=1+1)
_displayDemo/32-4      241µs ± 0%     230µs ± 0%   ~     (p=1.000 n=1+1)
_displayDemo/64-4      957µs ± 0%     920µs ± 0%   ~     (p=1.000 n=1+1)                                      
_displayDemo/128-4    3.85ms ± 0%    3.71ms ± 0%   ~     (p=1.000 n=1+1)                                        
_displayDemo/256-4    15.7ms ± 0%    15.1ms ± 0%   ~     (p=1.000 n=1+1)            
```
But worth more in other ilities like read- and flex-.