@jcorbin 2017-12-09
Yeah btw that strict 24 bit model? That's going to be a bigger perf boost
than almost anything else left; already peeked down that corridor, and the
map/if branch are quite noticeable in profiles; also fwiw, I wasn't able to
beat the map access with anything more ruthless (like a 256-case switch...)

Also there's a small bug in this PR that I'll push fix for once I'm not
roaming.
On Sat, Dec 9, 2017 at 12:38 Kris Kowal <notifications@github.com> wrote:

> *@kriskowal* commented on this pull request.
>
> I’ll take it.
>
> Aside: going over master, I noticed that we’re panic/asserting that the
> render model is the private mode type.
>
> We should either make the model struct (or func) public and use it in
> place of the interface…
>
> Or switch on the concrete type as an optimization for the case that it
> matches…
>
> I *do* expect that there will be further models, for example a strict 24
> bit color mode that does not fall back to palette colors, or a strict 256
> color mode that does not fall back to 4 or 5 bit colors (since some
> terminals remap the those, e.g., white to green or amber for nostalgia lols.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/borkshop/bork/pull/15#pullrequestreview-82334727>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AADCMzGBLT8t6iR8itmDE0jIoQdlmV4bks5s-u9JgaJpZM4Q8PEB>
> .
>