@jcorbin 2017-12-10
Model fix in #16.