@kriskowal 2017-12-10
Let’s just go for strict 24 by default. The fallbacks are optimizing for wire, at the expense of accurate color. The terminal is not guaranteed to respect the 256 color palette I’ve embedded.