@jcorbin 2017-12-10
You may be excited to know that this PR fixes a visual glitch in the braillewave demo, and I'm not really sure why...