@kriskowal 2017-12-10
@jcorbin Excited indeed! I couldn’t figure that out!